move : text 폴더 data 폴더로 이름 변경경
