+++ pages/contest/contest-detail.html
@@ -96,7 +96,7 @@
     </main>
     <footer></footer>
     <script src="/script/components/header.js"></script>
-    <script type="module" src="/script/text/contest-text.js"></script>
+    <script type="module" src="/script/data/contest-text.js"></script>
     <script type="module" src="/script/contest/contest-detail-render.js"></script>
   </body>
 </html>
