+++ pages/contest/contest.html
@@ -134,7 +134,7 @@
 
     <script src="/script/components/header.js"></script>
     <script src="/script/components/contest.js"></script>
-    <script type="module" src="/script/text/contest-text.js"></script>
+    <script type="module" src="/script/data/contest-text.js"></script>
     <script type="module" src="/script/contest/contest-list-render.js"></script>
   </body>
 </html>
