+++ pages/main/main.html
@@ -503,7 +503,7 @@
     <script src="../../script/main/total-search.js"></script>
     <script src="../../script/components/header.js"></script>
     <script src="../../script/components/contest.js"></script>
-    <script type="module" src="/script/text/contest-text.js"></script>
+    <script type="module" src="/script/data/contest-text.js"></script>
     <script type="module" src="../../script/main/contest-render.js"></script>
     <script src="../../script/components/license-card.js"></script>
     <script src="../../script/main/license-slider.js"></script>
