+++ script/data/contest-text.js
