+++ script/main/contest-render.js
@@ -1,4 +1,4 @@
-import { contestArr } from "../text/contest-text.js";
+import { contestArr } from "../data/contest-text.js";
 
 const contestSlider = () => {
   new Swiper(".contest-swiper", {
